Improved hooks.
Added 2 more hooks on a special rail that can transport the character to another part of the level and save a lot of time. Added also a hook with a box that can hit the player and can kill him when it passing through the rail on the 4 and 5 node.